Update Img table to include the image's file path instead of a blob
Reading many opinions from many developers, I found that storing the images in the file system is more efficient and a lot easier than storing them as a blob in the database. This is why the column image_path (varchar) will now hold the image's path, which can later also be a url or ip address to an image server.