Make the directory and file creation more dynamic
The methods that are responsible for finding a valid name for the images and created all necessary directories in the process are now more dynamic. The length of the names for subdirectories and the length of the names of images can now be changed using an environment variable or will be asked of the user at startup. This enables other people to use this web server more easily.